% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActiveCA-package.R
\docType{data}
\name{ActiveCA-package}
\alias{ActiveCA-package}
\title{ActiveCA: Processed data to obtain impedance functions for active transportation modes in Canada}
\description{
The main objective of the \emph{ActiveCA} package is to provide processed data to identify
the most appropriate impedance functions for active transportation modes considering
various destinations and time periods in Canada. To do achieve this objective,
we processed historical data from the \href{https://www23.statcan.gc.ca/imdb/p2SV.pl?Function=getSurvey&SDDS=5221}{General Social Survey (GSS)}
from 1986 to 2015 in Canada. We also calculate impedance function for cycling
and walking trips for the Census Agglomerations and Census Metropolitan Areas.
}
\references{
\url{https://github.com/dias-bruno/ActiveCA}
}
\author{
Bruno Dos Santos, School of Earth, Environment and Society, McMaster University \email{dossanb@mcmaster.ca}

Mahdis Moghadasi, School of Earth, Environment and Society, McMaster University \email{paezha@mcmaster.ca}

Antonio Paez, School of Earth, Environment and Society, McMaster University \email{paezha@mcmaster.ca}
}
