% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActiveCA-package.R
\docType{data}
\name{cycling_1998}
\alias{cycling_1998}
\title{Preprocessed cycling episodes (1998).}
\format{
A data frame containing 119 rows and 17 variables; each row represents a unique episode.
\describe{
\item{PUMFID}{Public Use Microdata File Identifier. This variable serves as the link between the Main and the Episode files. Each case has a unique record identifier and this appears on every episode}
\item{WGHT_EPI}{Episode weight. In 1998 the episode weight indicates the number of time use episodes that a record on the Episode File represents (see p. 9 of Cycle 12's Documentation and User's Guide).}
\item{ACTCODE}{Activity code of the episode.}
\item{act_label}{Activity (labeled).}
\item{STARTIME}{Start time (in minutes) of the episode.}
\item{ENDTIME}{End time (in minutes) of the episode.}
\item{DURATION}{Duration (in minutes) of the episode.}
\item{origin}{Place of origin of the episode (in code).}
\item{destination}{Place of destination of the episode (in code).}
\item{orig_label}{Place of origin of the episode (labeled).}
\item{dest_label}{Place of destination of the episode (labeled).}
\item{YEAR}{The year of the survey (1998).}
\item{MODE}{The transportation mode of the episode (cycling).}
\item{LUC_RST}{Code with the characteristic of the respondent's region of residence (how urbanized/rural).}
\item{PRV}{Code of the Province or Territory.}
\item{Pop_centre}{This variable informs if the respondent resides in a Census Metropolitan Area or Census Agglomeration.}
\item{Province}{Name of the Province or Territory.}
}
}
\source{
\href{https://www150.statcan.gc.ca/n1/pub/45-25-0001/index-eng.htm#a7}{General Social Survey (GSS)} / \href{http://odesi2.scholarsportal.info/webview/}{Odesi}

Adapted from Statistics Canada, General Social Survey Cycle 12, 1998. This does not constitute an endorsement by Statistics Canada of this product.
}
\usage{
data(cycling_1998)
}
\description{
This dataset contains all the cycling episodes from the 1998 GSS Survey.
We identified the cycling episodes and selected the places of origin
and destination, measuring the time spent, weight and other
information about the activities.
}
\seealso{
Cycle 12: Time Use. Public Use Microdata File \href{https://publications.gc.ca/collections/collection_2017/statcan/12-0012/CS12-0012-1999-eng.pdf}{Documentation and User’s Guide}
}
\keyword{1998}
\keyword{cycling}
