% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActiveCA-package.R, R/_ActiveCA-package.R
\docType{data}
\name{gss_episodes}
\alias{gss_episodes}
\title{Preprocessed walking and cycling episodes (from 1986 to 2015).}
\format{
A data frame containing 21748 rows and 17 variables; each row represents a unique episode.
\describe{
\item{PUMFID}{Uniquely identifies the episode. This variable serves as the link between the Main and the Episode files. Each case has a unique record identifier and this appears on every episode}
\item{WGHT_EPI}{Episode weight.}
\item{ACTCODE}{Activity code of the episode.}
\item{STARTIME}{Start time (in minutes) of the episode.}
\item{ENDTIME}{End time (in minutes) of the episode.}
\item{DURATION}{Duration (in minutes) of the episode.}
\item{LOCATION}{Location/mode of transport of the episode.}
\item{origin}{Place of origin of the episode (in code).}
\item{destination}{Place of destination of the episode (in code).}
\item{orig_label}{Place of origin of the episode (labeled).}
\item{dest_label}{Place of destination of the episode (labeled).}
\item{YEAR}{The year of the survey.}
\item{MODE}{The transportation mode of the episode (walking or cycling).}
\item{LUC_RST}{Code with the characteristic of the respondent's region of residence (how urbanized/rural).}
\item{PRV}{Code of the Province or Territory.}
\item{Pop_centre}{This variable informs if the respondent resides in a Census Metropolitan Area or Census Agglomeration.}
\item{Province}{Name of the Province or Territory.}
}

A data frame containing 21748 rows and 17 variables; each row represents a unique episode.
\describe{
\item{PUMFID}{Uniquely identifies the episode. This variable serves as the link between the Main and the Episode files. Each case has a unique record identifier and this appears on every episode}
\item{WGHT_EPI}{Episode weight.}
\item{ACTCODE}{Activity code of the episode.}
\item{STARTIME}{Start time (in minutes) of the episode.}
\item{ENDTIME}{End time (in minutes) of the episode.}
\item{DURATION}{Duration (in minutes) of the episode.}
\item{LOCATION}{Location/mode of transport of the episode.}
\item{origin}{Place of origin of the episode (in code).}
\item{destination}{Place of destination of the episode (in code).}
\item{orig_label}{Place of origin of the episode (labeled).}
\item{dest_label}{Place of destination of the episode (labeled).}
\item{YEAR}{The year of the survey.}
\item{MODE}{The transportation mode of the episode (walking or cycling).}
\item{LUC_RST}{Code with the characteristic of the respondent's region of residence (how urbanized/rural).}
\item{PRV}{Code of the Province or Territory.}
\item{Pop_centre}{This variable informs if the respondent resides in a Census Metropolitan Area or Census Agglomeration.}
\item{Province}{Name of the Province or Territory.}
}
}
\source{
"\href{https://www23.statcan.gc.ca/imdb/p2SV.pl?Function=getSurvey&SDDS=5221}{General Social Survey (GSS)} / \href{http://odesi2.scholarsportal.info/webview/}{Odesi}"

"\href{https://www23.statcan.gc.ca/imdb/p2SV.pl?Function=getSurvey&SDDS=5221}{General Social Survey (GSS)} / \href{http://odesi2.scholarsportal.info/webview/}{Odesi}"
}
\usage{
data(gss_episodes)

data(gss_episodes)
}
\description{
This dataset contains all the walking and cycling episodes from the 1986 to 2015 GSS Surveys.
We identified the cycling episodes and selected the places of origin
and destination, measuring the time spent, weight and other
information about the activities.

This dataset contains all the walking and cycling episodes from the 1986 to 2015 GSS Surveys.
We identified the cycling episodes and selected the places of origin
and destination, measuring the time spent, weight and other
information about the activities.
}
\keyword{cycling}
\keyword{episodes}
\keyword{gss}
\keyword{walking}
