% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActiveCA-package.R
\docType{data}
\name{gss_impedances}
\alias{gss_impedances}
\title{Impedance functions for Census Agglomerations and Census Metropolitan Areas (from 1986 to 2022).}
\format{
A data frame containing 13585 rows and 19 variables; each row represents a unique episode.
\describe{
\item{PUMFID}{Public Use Microdata File Identifier. This variable serves as the link between the Main and the Episode files. Each case has a unique record identifier and this appears on every episode}
\item{WGHT_EPI}{Episode weight.}
\item{DURATION}{Duration (in minutes) of the episode.}
\item{origin}{Place of origin of the episode (in code).}
\item{destination}{Place of destination of the episode (in code).}
\item{Province}{Name of the Province or Territory.}
\item{Pop_centre}{This variable informs if the respondent resides in a Census Metropolitan Area or Census Agglomeration.}
\item{orig_label}{Place of origin of the episode (labeled).}
\item{dest_label}{Place of destination of the episode (labeled).}
\item{YEAR}{The year of the survey.}
\item{MODE}{The transportation mode of the episode (walking or cycling).}
\item{f_name}{Name of the function that best adjusted the time values for the combination of year, destination and mode (can be “lnorm”, “gamma” or “unif”)}.
\item{f}{Impedance value obtained for the duration time after applying the function described in fname.}
\item{est_1}{Estimate of the first parameter for the function described in fname.}
\item{est_2}{Estimate of the second parameter for the function described in fname.}
\item{loglik}{Log-Likelihood obtained for the function fname.}
\item{aic}{Akaike information criterion obtained for function fname.}
\item{bic}{Bayesian information criterion obtained for function fname.}
\item{count}{Count of year-destination-mode combinations used to obtain the impedance function.}
}
}
\source{
\href{https://www23.statcan.gc.ca/imdb/p2SV.pl?Function=getSurvey&SDDS=5221}{General Social Survey (GSS)} / \href{http://odesi2.scholarsportal.info/webview/}{Odesi}
}
\usage{
data(gss_impedances)
}
\description{
This dataset contains all impedance functions for walking
and cycling transportation, obtained from the 1986 to 2022
GSS Surveys, for Census Metropolitan Areas and Census Agglomerations.
We estimated the best impedance function for each combination of year,
destination and transportation mode, using the fitdistrplus package.
The best-fit function was defined by selecting the lowest Akaike information criterion.
}
\keyword{cycling}
\keyword{gss}
\keyword{impedance}
\keyword{walking}
