% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActiveCA-package.R
\docType{data}
\name{gss_main_1986}
\alias{gss_main_1986}
\title{Selection of socioeconomic variables from the 1986 GSS Time Use Survey (Cycle 2).}
\format{
A data frame containing 16390 rows and 9 variables; each row represents a unique respondent.
\describe{
\item{SEQNUM}{Record identification.}
\item{FWGT_OS}{Final Weight for Main File.}
\item{DVAGEGR}{Age group of respondent.}
\item{DVSEX}{Sex of the respondent.}
\item{DVMARST}{Marital status of the respondent.}
\item{DVHHLDSZ}{Household size.}
\item{PROV}{Province of residence of the respondent.}
\item{DVCHILD}{Number of children.}
}
}
\source{
\href{https://www23.statcan.gc.ca/imdb/p2SV.pl?Function=getSurvey&SDDS=5221}{General Social Survey (GSS)} / \href{http://odesi2.scholarsportal.info/webview/}{Odesi}
}
\usage{
data(gss_main_1986)
}
\description{
This dataset contains some socioeconomic variables obtained
from the main file of the 1986 GSS Time Use Survey (Cycle 2). We labeled
variables according to the Public Use Micro Data File Documentation and User's Guide (gss-12M0002-E-1986-c-2-main).
}
\keyword{1986}
\keyword{economic}
\keyword{gss}
\keyword{main}
\keyword{respondent}
\keyword{socio}
