% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActiveCA-package.R
\docType{data}
\name{gss_main_1992}
\alias{gss_main_1992}
\title{Selection of socioeconomic variables from the 1992 GSS Time Use Survey (Cycle 7).}
\format{
A data frame containing 9815 rows and 9 variables; each row represents a unique respondent.
\describe{
\item{SEQNUM}{Record identification.}
\item{FWGHT}{Final Weight for Main File.}
\item{SURMNTH}{Survey Month of Questionnaire.}
\item{FORMTYPE}{Form Type - English or French.}
\item{A4}{How often do you feel rushed? Would you say it is.}
\item{A5}{Compared to 5 years ago, do you feel more rushed?.}
\item{A6}{How often do you feel you have time on your hands?.}
\item{C2}{Respondent has child(ren) <15?.}
\item{DVAGEGR}{Age group of respondent.}
\item{DVSEX}{Sex of respondent.}
\item{DVMS}{Marital status of respondent.}
\item{DVHHSCAP}{Total number of persons living in the household.}
\item{DVPROV}{Province of residence of respondent.}
\item{DVCMA}{Derived Census metropolitan area.}
\item{DVLVGRGR}{Grouped living arrangement of the respondent.}
\item{DVAGRYC}{Age of respondents youngest single child.}
\item{RESC0004}{Number of respondents single children aged 0-4.}
\item{RESC0512}{Number of respondents single children aged 5-12.}
\item{RESC1318}{Number of respondents single children aged 13-18.}
\item{RESC1924}{Number of respondents single children aged 19-24.}
\item{RESC2596}{Number of respondents single children aged 25+.}
\item{DVAGRSP}{Age group of the respondents spouse.}
}
}
\source{
\href{https://www23.statcan.gc.ca/imdb/p2SV.pl?Function=getSurvey&SDDS=5221}{General Social Survey (GSS)} / \href{http://odesi2.scholarsportal.info/webview/}{Odesi}
}
\usage{
data(gss_main_1992)
}
\description{
This dataset contains some socioeconomic variables obtained
from the main file of the 1992 GSS Time Use Survey (Cycle 7). We labeled
variables according to the Public Use Micro Data File Documentation and User's Guide (gss-12M0007-E-1992-c-7-m).
}
\keyword{1992}
\keyword{economic}
\keyword{gss}
\keyword{main}
\keyword{respondent}
\keyword{socio}
