% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActiveCA-package.R
\docType{data}
\name{gss_main_1998}
\alias{gss_main_1998}
\title{Selection of socioeconomic variables from the 1998 GSS Time Use Survey (Cycle 12).}
\format{
A data frame containing 10749 rows and 22 variables; each row represents a unique respondent.
\describe{
\item{RECID}{Record identification.}
\item{WGHTFIN}{Final Person Weight.}
\item{SURVMNTH}{Survey month of data collection.}
\item{LANINT}{Language of interview.}
\item{AGEC}{Age of respondent.}
\item{AGEGR5}{Age group of respondent in group of 5.}
\item{AGEGR10}{Age group of respondent in group of 10.}
\item{SEX}{Sex of respondent.}
\item{MARSTAT}{Marital status of respondent.}
\item{AGEPRGR5}{Age group of respondent's spouse/partner.}
\item{PRTYPEC}{Type of partner the respondent has within the household.}
\item{CHRFLAG}{Child(ren) of the respondent living in the household.}
\item{AGECHRYC}{Age of respondent's youngest single child living in the household.}
\item{CHR0004C}{Number of respondent's children 0 to 4 years living in the household.}
\item{CHR0512C}{Number of respondent's children 5 to 12 years living in the household.}
\item{CHR1314C}{Number of respondent's children 13 to 14 years living in the household.}
\item{CHR1518C}{Number of respondent's children 15 to 18 years living in the household.}
\item{CHR1924C}{Number of respondent's children 19 to 24 years living in the household.}
\item{CHR25UPC}{Number of respondent's children 25 years and older years living in the household.}
\item{CHH0004C}{Number of children aged from 0 to 4 living in respondent's household.}
\item{CHH0512C}{Number of children aged from 5 to 12 living in respondent's household.}
\item{CHH1314C}{Number of children aged from 5 to 12 living in respondent's household.}
\item{CHRTIME6}{Number of respondents children living at home.}
\item{PARHSD}{Type of parents the respondent has within the household.}
\item{LIVARR12}{Living arrangement of respondent's household (in group of 12).}
\item{LIVARR08}{Living arrangement of respondent's household (in group of 8).}
\item{HSDSIZEC}{Household size of respondent.}
\item{FAMTYPE}{Respondent's type of family structure.}
\item{MULTIGEN}{Three-generation family in respondent's household.}
\item{PRV}{Province of residence of the respondent.}
\item{REGION}{Region of residence of the respondent.}
\item{CMAPRV}{Census Metropolitan Area (CMA) of the respondent.}
\item{A2}{How often do you feel rushed?.}
\item{A3}{Compared to five years ago, do you feel.}
\item{A4}{How often do you feel you have time on your hands.}
\item{A5}{Do you feel that weekdays are just too short.}
\item{A6}{On which main activity would you choose to spend more time on.}
\item{DDAY}{Designated day of interview.}
\item{DIARYDAY}{Date of interview.}
\item{DIARYREC}{Recall of diary interview.}
\item{DVTDAY}{Type of day of interview.}
}
}
\source{
\href{https://www23.statcan.gc.ca/imdb/p2SV.pl?Function=getSurvey&SDDS=5221}{General Social Survey (GSS)} / \href{http://odesi2.scholarsportal.info/webview/}{Odesi}
}
\usage{
data(gss_main_1998)
}
\description{
This dataset contains some socioeconomic variables obtained
from the main file of the 1998 GSS Time Use Survey (Cycle 12). We labeled
variables according to the Public Use Micro Data File Documentation and User's Guide (gss-12M0012-E-1998-c-12-m).
}
\keyword{1998}
\keyword{economic}
\keyword{gss}
\keyword{main}
\keyword{respondent}
\keyword{socio}
