% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActiveCA-package.R
\docType{data}
\name{gss_main_2005}
\alias{gss_main_2005}
\title{Selection of socioeconomic variables from the 2005 GSS Time Use Survey (Cycle 19).}
\format{
A data frame containing 19597 rows and 32 variables; each row represents a unique respondent.
\describe{
\item{RECID}{Record identification.}
\item{WGHT_PER}{Person weight.}
\item{WGHT_HSD}{Household weight.}
\item{WGHT_CSP}{Culture, sport and physical activity sample weight.}
\item{WGHT_SNT}{Social networks sample weight.}
\item{AGEGR5}{Age group of the respondent (group of 5).}
\item{AGEGR10}{Age group of the respondent (group of 10).}
\item{SEX}{Sex of respondent.}
\item{MARSTAT}{Marital status of the respondent.}
\item{AGEPRGRDIF}{Age difference between respondent and spouse/partner.}
\item{SEXPR}{Sex of the respondent's spouse/partner within the household.}
\item{PRTYPEC}{Type of partner the respondent has within the household.}
\item{AGECHRYC}{Age of respondent's youngest single child living in the household.}
\item{CHRFLAG}{Single child(ren) of the respondent living in the household.}
\item{CHRINHSDC}{Number of respondent's child(ren) living in the household(any age or marital status).}
\item{CHR0014C}{Number of respondent's child(ren) 0 to 14 years of age living in the household.}
\item{CHRTIME6}{Number of respondent's child(ren) living at home.}
\item{PARHSDC}{Type of parents the respondent has within the household.}
\item{LIVARR08}{Living arrangement of respondent's household (in group of 12)}
\item{LIVARR12}{Living arrangement of respondent's household (in group of 8)}
\item{HSDSIZEC}{Household size of respondent.}
\item{FAMTYPE}{Respondent's type of family structure.}
\item{MULTIGEN}{Three-generation or more family in respondent's household.}
\item{PRV}{Province of residence of the respondent.}
\item{REGION}{Region of residence of the respondent.}
\item{LUC_RST}{Urban/Rural indicator.}
\item{GTU_Q110}{How often do you feel rushed? Would you say it is:}
\item{GTU_Q120}{Compared to five years ago, do you feel more rushed, about the same or less rushed?}
\item{GTU_Q130}{How often do you feel you have time on your hands that you Don't know what to do with? Would you say it is:}
\item{GTU_Q140}{Do you feel that the days are just too short to do all the things you want?}
\item{GTU_Q150}{On which main activity would you choose to spend more time if you could?}
\item{DVTDAY}{Type of day of the designated day of interview.}
}
}
\source{
\href{https://www23.statcan.gc.ca/imdb/p2SV.pl?Function=getSurvey&SDDS=5221}{General Social Survey (GSS)} / \href{http://odesi2.scholarsportal.info/webview/}{Odesi}
}
\usage{
data(gss_main_2005)
}
\description{
This dataset contains some socioeconomic variables obtained
from the main file of the 2005 GSS Time Use Survey (Cycle 19). We labeled
variables according to the Public Use Micro Data File Documentation and User's Guide (Product No.12M0019-GPE).
}
\keyword{2005}
\keyword{economic}
\keyword{gss}
\keyword{main}
\keyword{respondent}
\keyword{socio}
