% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActiveCA-package.R
\docType{data}
\name{gss_main_2015}
\alias{gss_main_2015}
\title{Selection of socioeconomic variables from the 2015 GSS Time Use Survey (Cycle 24).}
\format{
A data frame containing 15390 rows and 33 variables; each row represents a unique respondent.
\describe{
\item{PUMFID}{Record identification.}
\item{WGHT_PER}{Person weight.}
\item{SURVMNTH}{Survey month of data collection.}
\item{AGEGR10}{Age group of the respondent (group of 10).}
\item{SEX}{Sex of respondent.}
\item{MARSTAT}{Marital status of the respondent.}
\item{PHSDFLG}{Respondent has a spouse/partner living in the household.}
\item{AGEPRGRD}{Age difference between respondent and spouse/partner.}
\item{SEXPR}{Sex of the respondent's spouse/partner within the household.}
\item{PRTYPEC}{Sex of the respondent's spouse/partner within the household.}
\item{CXRFLAG}{Type of partner the respondent has within the household.}
\item{AGECXRYG}{Age of respondent’s youngest child in household.}
\item{AGEHSDYC}{Age of youngest household member in respondent’s household.}
\item{CHINHSDC}{Respondent’s child(ren) in household - Any age/marital status.}
\item{CXR0014C}{Respondent’s child(ren) in household - 0 to 14 years.}
\item{CXRTIME6}{Age group of respondent’s child(ren) in household.}
\item{CHH0014C}{Child(ren) in household - 0 to 14 years.}
\item{PARNUM}{Number of parents the respondent has in household.}
\item{LIVARR08}{Living arrangement of respondent's household (in group of 12).}
\item{LIVARR11}{Living arrangement of respondent's household (in group of 8).}
\item{HSDSIZEC}{Household size of respondent.}
\item{HSDELIGC}{Number of members in respondent’s household 15 years of age or older.}
\item{MULTIGEN}{Three generations or more family in respondent's household.}
\item{PRV}{Province of residence of the respondent.}
\item{LUC_RST}{Population centres indicator.}
\item{GTU_110}{How often do you feel rushed? Would you say it is:.}
\item{GTU_130}{How often do you feel you have time on your hands that you Don't know what to do with? Would you say it is:.}
\item{DVTDAY}{Type of day of the designated day of interview.}
\item{SLM_01}{Using a scale of 0 to 10 where 0 means "Very dissatisfied" and 10 means "Very satisfied", how do you feel about your life as a whole right now?}
\item{SRH_110}{In general, would you say your health is...?}
\item{SRH_115}{In general, would you say your mental health is...?}
}
}
\source{
\href{https://www23.statcan.gc.ca/imdb/p2SV.pl?Function=getSurvey&SDDS=5221}{General Social Survey (GSS)} / \href{http://odesi2.scholarsportal.info/webview/}{Odesi}
}
\usage{
data(gss_main_2015)
}
\description{
This dataset contains some socioeconomic variables obtained
from the main file of the 2015 GSS Time Use Survey (Cycle 24). We labeled
variables according to the Public Use Micro Data File Documentation and User's Guide (Catalogue no. 12M0024X).
}
\keyword{2015}
\keyword{economic}
\keyword{gss}
\keyword{main}
\keyword{respondent}
\keyword{socio}
