% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActiveCA-package.R
\docType{data}
\name{gss_main_2022}
\alias{gss_main_2022}
\title{Selection of socioeconomic variables from the 2022 GSS Time Use Survey.}
\format{
A data frame containing 12,336 rows and 101 variables; each row represents a unique respondent.
\describe{
\item{PUMFID}{Record identification.}
\item{WGHT_PER}{Person weight.}
\item{SURVMNTH}{Survey month of data collection.}
\item{PRV}{Province of residence of the respondent.}
\item{LUC_RST}{Population centres indicator.}
\item{REGION}{Region of residence of the respondent.}
\item{HSDSIZEC}{Household size of respondent.}
\item{AGEGR10}{Age group of the respondent (group of 10).}
\item{GENDER2}{Gender (2) after distribution of the non-binary persons (Men+ or Women+).}
\item{MARSTAT}{Marital status of the respondent.}
\item{PHSDFLG}{Respondent has a spouse/partner living in the household.}
\item{AGEPRC}{Age group of respondent’s spouse/partner.}
\item{GENPR2}{Gender of spouse/partner after distribution of the non-binary person (Men+ or Women+).}
\item{COUPGEN2}{Gender+ of household couple.}
\item{CHH0017C}{Number of child(ren) in household - 0 to 17 years (capped).}
\item{CHH0004C}{Child(ren) in household - 0 to 4 years (flag).}
\item{CHH0514C}{Child(ren) in household - 5 to 14 years (flag).}
\item{CHH1517C}{Child(ren) in household - 15-17 years (flag).}
\item{CXRFLAG}{Child(ren) of respondent living in household.}
\item{PARNUM}{Number of parents the respondent has in household.}
\item{SENFLAG}{Senior(s) (age 65 and over) in household.}
\item{MULTIGEN}{Three generations or more in the respondent’s household.}
\item{DDEV_FL}{Disability indicator - Developmental.}
\item{DDEX_FL}{Disability indicator - Dexterity.}
\item{DFLEX_FL}{Disability indicator - Flexibility.}
\item{DHEAR_FL}{Disability indicator - Hearing.}
\item{DLRN_FL}{Disability indicator - LearninG.}
\item{DMEM_FL}{Disability indicator - Memory.}
\item{DMENT_FL}{Disability indicator - Mental health-related.}
\item{DMOB_FL}{Disability indicator - Mobility.}
\item{DPAIN_FL}{Disability indicator - Pain-related.}
\item{DUNK_FL}{Disability indicator - Unknown disability.}
\item{DVIS_FL}{Disability indicator - Seeing.}
\item{DDIS_FL}{Disability status.}
\item{DTYPER}{Disability type counter - Grouped.}
\item{DCLASS}{Global (disability) severity class.}
\item{TCS_110}{Perceptions of time - Plans to slow down. Do you plan to slow down in the coming year?}
\item{TCS_120}{Perceptions of time - Workaholic. Do you consider yourself a workaholic?}
\item{TCS_130}{Perceptions of time - Tends to cut back on sleep. When you need more time, do you tend to cut back on your sleep?}
\item{TCS_140}{Perceptions of time - Not accomplishing what you set out to do. At the end of the day, do you often feel that you have not accomplished what you had set out to do due to lack of time?}
\item{TCS_150}{Perceptions of time - Not spending enough time with family or friends. Do you worry that you don’t spend enough time with your family or friends?}
\item{TCS_160}{Perceptions of time - Constantly under stress. Do you feel that you’re constantly under stress trying to accomplish more than you can handle?}
\item{TCS_170}{Perceptions of time - Trapped in daily routine. Do you feel trapped in a daily routine?}
\item{TCS_180}{Perceptions of time - No time for fun. Do you feel that you just don’t have time for fun any more?}
\item{TCS_190}{Perceptions of time - Stress when there is not enough time. Do you often feel under stress when you don’t have enough time?}
\item{TCS_200}{Perceptions of time - Would like more time alone. Would you like to spend more time alone?}
\item{TIMECR}{Time crunch indicator. This derived variable measures the number of "Yes" codes reported in the questions TCS_Q110 to TCS_Q200.}
\item{GEN_01}{Self rated health.}
\item{GEN_02}{Self rated mental health.}
\item{SRS_10}{Self Rated Stress - Amount.}
\item{MSS_130}{What is your main source of stress?.}
\item{ACT7DAYC}{Main activity - Last week (collapsed).}
\item{MRW_05C}{Main activity - Last 12 months (collapsed).}
\item{MRW_D40A}{Worked in the last 12 months.}
\item{MRW_D40B}{Worked last week.}
\item{WET_120}{Job type during weeks employed.}
\item{WHW_230}{Usual work schedule at main job.}
\item{WHWD140G}{Hours worked at all jobs in a week (grouped).}
\item{TLWK_01A}{Telework last week - From home.}
\item{TLWK_06}{Telework aspects - Avoid the commute. How important is the opportunity not to commute every day to your workplace?}
\item{WLYD170G}{Distance from residence to workplace (grouped).}
\item{ATT_120}{Vehicle at respondent’s disposal.}
\item{ATT_130}{Need someone else to drive - Frequency.}
\item{ATT_131}{Someone else can drive when needed - Frequency.}
\item{ATT_150C}{Most common kind of transportation - Typical week (collapsed).}
\item{ATT_160}{Access to public transportation.}
\item{ATT_170}{Use of public transportation - Past 12 months.}
\item{CTW_140A}{Commute to work/school - Car, truck or van - As driver.}
\item{CTW_140B}{Commute to work/school - Car, truck or van - As passenger.}
\item{CTW_140C}{Commute to work/school - Public transit.}
\item{CTW_140D}{Commute to work/school - Walked.}
\item{CTW_140E}{Commute to work/school - By bicycle.}
\item{CTW140GR}{Commute to work/school - Other (grouped).}
\item{CTW_140I}{Commute to work/school - Worked/Attended school at home.}
\item{CTW_150G}{Commute to work/school - Length of time one way (grouped).}
\item{CTW_160}{Commute to work/school - Ever used public transit.}
\item{CTW_170}{Commute to work/school - Convenience of public transit (user).}
\item{CTW_180}{Commute to work/school - Convenience of public transit (non-user).}
\item{CTW_190}{Commute to work/school - Traffic congestion (frequency).}
\item{CTW_210}{Commute to work/school - Satisfaction with commute time.}
\item{CTW_215}{Commute to work - Traffic congestion (seriousness).}
\item{ED_05}{Educational attainment - Highest degree (7 categories).}
\item{MAP_110C}{Main activity - Spouse/partner - 12 months (collapsed).}
\item{MAP_D40A}{Spouse/partner - worked in the last 12 months.}
\item{DOR_110C}{Dwelling type (collapsed) .}
\item{IM_01A}{Born in/outside Canada - Respondent.}
\item{YRARRI_C}{Year respondent first came to Canada to live (collapsed).}
\item{BPP_01A}{Place of birth of respondent’s spouse/partner - Canada.}
\item{BPR_20}{Parents born in or outside of Canada.}
\item{ABM_01A}{Indigenous Identity - Not First Nations, Métis or Inuk (Inuit).}
\item{AIP_01A}{Indigenous identity of spouse/partner - Not Indigenous.}
\item{VISMINFL}{Visible minority flag.}
\item{VISPRFL}{Visible minority status - Respondent’s spouse/partner.}
\item{REL_02}{Frequency of religious participation - Past 12 months.}
\item{RELIGFLG}{Religious affiliation flag.}
\item{INC_C}{Income - Total (before tax) (grouped).}
\item{FAMINC_C}{Family income - Total (before tax) (grouped).}
\item{GTU_110}{General time use - Feel rushed. How often do you feel rushed?}
\item{GTU_130}{General time use - Extra time. How often do you feel you have time on your hands that you don’t know what to do with?}
\item{DDAY}{Diary - Reference day.}
}
}
\source{
\href{https://www23.statcan.gc.ca/imdb/p2SV.pl?Function=getSurvey&SDDS=5221}{General Social Survey (GSS)} / \href{http://odesi2.scholarsportal.info/webview/}{Odesi}
}
\usage{
data(gss_main_2022)
}
\description{
This dataset contains some socioeconomic variables obtained
from the main file of the 2022 GSS Time Use Survey. We labeled
variables according to the Public Use Micro Data File Documentation and User's Guide (General Social Survey (Time Use) 2022).
}
\keyword{2022}
\keyword{economic}
\keyword{gss}
\keyword{main}
\keyword{respondent}
\keyword{socio}
