% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActiveCA-package.R
\docType{package}
\name{ActiveCA-package}
\alias{ActiveCA}
\alias{ActiveCA-package}
\title{ActiveCA: Time Use Data from the General Social Survey of Canada to Study Active Travel}
\description{
ActiveCA is an Open Data Product that provides analysis-ready data from the General Social Survey of Canada Cycles 2 (1986), 7 (1992), 12 (1998), 19 (2005), 24 (2010), 29 (2015), and 34 (2022) regarding active travel. These data can be used to study walking and cycling in Canada for various purposes over time, and to calculate impedance functions for accessibility analysis. The package provides pre-estimated impedance functions for cycling and walking trips for Census Agglomerations and Census Metropolitan Areas.
}
\author{
\strong{Maintainer}: Bruno Dos Santos \email{dossanb@mcmaster.ca} (\href{https://orcid.org/0000-0002-6748-2038}{ORCID})

Authors:
\itemize{
  \item Mahdis Moghadasi \email{moghadasi.mahdis@mcmaster.ca} (\href{https://orcid.org/0000-002-6748-2038}{ORCID})
  \item Antonio Paez \email{paezha@mcmaster.ca} (\href{https://orcid.org/0000-0001-6912-9919}{ORCID})
}

Other contributors:
\itemize{
  \item Statistics Canada (Copyright holder of the Public Use Microdata Files of the General Social Survey. These files are released using Statistics Canada Open Licence) [copyright holder]
}

}
\keyword{internal}
