% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActiveCA-package.R, R/_ActiveCA-package.R
\docType{package}
\name{ActiveCA-package}
\alias{ActiveCA-package}
\title{ActiveCA: Time Use Data from the General Social Survey of Canada to Study Active Travel}
\description{
ActiveCA is an Open Data Product that provides analysis-ready data from the General Social Survey of Canada Cycles XX (YEAR), XX (YEAR), XX (YEAR), XX (YEAR), and XX (YEAR) regarding active travel. These data can be used to study walking and cycling in Canada for various purposes over time, and to calculate impedance functions for accessibility analysis. The package provides pre-estimated impedance functions for cycling and walking trips for Census Agglomerations and Census Metropolitan Areas.

The main objective of the \emph{ActiveCA} package is to provide processed data to identify
the most appropriate impedance functions for active transportation modes considering
various destinations and time periods in Canada. To do achieve this objective,
we processed historical data from the \href{https://www23.statcan.gc.ca/imdb/p2SV.pl?Function=getSurvey&SDDS=5221}{General Social Survey (GSS)}
from 1986 to 2015 in Canada. We also calculate impedance function for cycling
and walking trips for the Census Agglomerations and Census Metropolitan Areas.
}
\references{
\url{https://github.com/dias-bruno/ActiveCA}
}
\author{
\strong{Maintainer}: Bruno Dos Santos \email{dossanb@mcmaster.ca} (\href{https://orcid.org/0000-0002-6748-2038}{ORCID})

Authors:
\itemize{
  \item Mahdis Moghadasi \email{moghadasi.mahdis@mcmaster.ca} (\href{https://orcid.org/0000-002-6748-2038}{ORCID})
  \item Antonio Paez \email{paezha@mcmaster.ca} (\href{https://orcid.org/0000-0001-6912-9919}{ORCID})
}


Bruno Dos Santos, School of Earth, Environment and Society, McMaster University \email{dossanb@mcmaster.ca}

Mahdis Moghadasi, School of Earth, Environment and Society, McMaster University \email{paezha@mcmaster.ca}

Antonio Paez, School of Earth, Environment and Society, McMaster University \email{paezha@mcmaster.ca}
}
\keyword{internal}
